Выполнила задание 5.14. Модуляция
