+++ js/data.js
@@ -1,0 +1,99 @@
+import { getRandomInteger, getRandomArrayElement } from './utils';
+
+const POSTS_COUNT = 25;
+
+const NAMES = [
+  'Angela',
+  'Mark',
+  'Tim',
+  'Piter',
+  'Jimmy',
+  'Laura',
+  'Kelly',
+  'Ingrid',
+  'Richard',
+  'Roland',
+  'Hank',
+  'Edward',
+  'Anna',
+  'Maria',
+  'Diana',
+  'Will',
+  'Bill',
+  'John',
+  'Brooke',
+  'Shaila',
+  'Celine',
+  'Kirsten',
+  'Paul',
+  'Otis',
+  'Lenny',
+];
+
+const DESCRIPTIONS = [
+  'С учётом сложившейся международной обстановки, существующая теория предполагает независимые способы реализации экономической целесообразности принимаемых решений.',
+  'Лишь предприниматели в сети интернет обнародованы.',
+  'Есть над чем задуматься: тщательные исследования конкурентов представляют собой не что иное, как квинтэссенцию победы маркетинга над разумом и должны быть обнародованы.',
+  'В частности, современная методология разработки обеспечивает актуальность соответствующих условий активизации.',
+  'Господа, экономическая повестка сегодняшнего дня выявляет срочную потребность поэтапного и последовательного развития общества.',
+  'Есть над чем задуматься: предприниматели в сети интернет лишь добавляют фракционных разногласий и указаны как претенденты на роль ключевых факторов.',
+  'В своём стремлении повысить качество жизни, они забывают, что глубокий уровень погружения в значительной степени обусловливает важность поэтапного и последовательного развития общества.',
+  'Не следует, однако, забывать, что синтетическое тестирование прекрасно подходит для реализации форм воздействия.',
+  'Внезапно, предприниматели в сети интернет освещают чрезвычайно интересные особенности картины в целом, однако конкретные выводы, разумеется, ассоциативно распределены по отраслям.',
+  'Картельные сговоры не допускают ситуации, при которой представители современных социальных резервов разоблачены. ',
+  'Банальные, но неопровержимые выводы, а также многие известные личности будут ограничены исключительно образом мышления.',
+  'Прежде всего, курс на социально-ориентированный национальный проект предоставляет широкие возможности для поставленных обществом задач. ',
+  'C помощью этого онлайн-генератора рыботекста можно пачками плодить как отдельные предложения и заголовки, так и целые абзацы отменнейшего рыбы-текста.',
+  'Прародителем текста-рыбы является известный "Lorem Ipsum" — латинский текст, ноги которого растут аж из 45 года до нашей эры. ',
+  'Для формирования текста комментария — message — вам необходимо взять одно или два случайных предложения из представленных ниже',
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const likesCount = {
+  MIN: 15,
+  MAX: 200,
+};
+
+const avatarCount = {
+  MIN: 1,
+  MAX: 6,
+};
+
+const commentsQuantity = {
+  MIN: 0,
+  MAX: 30,
+};
+
+const createComment = (id) => ({
+  id,
+  avatar: `img/avatar-${getRandomInteger(
+    avatarCount.MIN,
+    avatarCount.MAX
+  )}.svg`,
+  message: getRandomArrayElement(MESSAGES),
+  name: getRandomArrayElement(NAMES),
+});
+
+const createPost = (id) => ({
+  id,
+  url: `photos/${id}.jpg`,
+  description: getRandomArrayElement(DESCRIPTIONS),
+  likes: getRandomInteger(likesCount.MIN, likesCount.MAX),
+  comments: Array.from(
+    { length: getRandomInteger(commentsQuantity.MIN, commentsQuantity.MAX) },
+    (_, index) => createComment(index + 1)
+  ),
+});
+
+const createPosts = () =>
+  Array.from({ length: POSTS_COUNT }, (_, index) => createPost(index + 1));
+
+export { createPosts };
